requirements.txt
python-dotenv
requests
fastapi
uvicorn[standard]